refactor: Mudança de Session.duration_hours para Numeric(10,4) e ajuste de cálculos

+++ app/models/session.py
@@ -1,9 +1,10 @@
 # app/models/session.py
 
-from sqlalchemy import Column, Integer, ForeignKey, Float, DateTime
+from sqlalchemy import Column, Integer, ForeignKey, Float, DateTime, Numeric
 from sqlalchemy.orm import relationship
 from app.database import Base
 from sqlalchemy.sql import func
+
 
 class Session(Base):
     __tablename__ = "sessions"
@@ -11,7 +12,8 @@ class Session(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"))
     goal_id = Column(Integer, ForeignKey("goals.id", ondelete="SET NULL"))
-    duration_hours = Column(Float, nullable=False)
+    # precisão de horas com 4 casas decimais
+    duration_hours = Column(Numeric(10, 4), nullable=False)
 
     started_at = Column(DateTime(timezone=True), server_default=func.now())
     finished_at = Column(DateTime(timezone=True), nullable=True)
